requirements.txt

colorama==0.4.6
python-dotenv==1.0.1
twine!=5.1.0
build<0.10.0
awswrangler==3.8.0
pandas==2.2.3
numpy==1.26.4
boto3==1.34.48
oracledb==2.2.1
cx-Oracle==8.3.0
pyodbc==5.1.0
psycopg2==2.9.9
pyspark==3.5.1
redshift_connector==2.1.7
Requests==2.32.3
SQLAlchemy==2.0.31
google-cloud-aiplatform==1.63.0
google-cloud-storage==2.19.0
google_cloud_bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
sqlalchemy-bigquery==1.11.0